Move functions from language.specification to LanguageGraph
