+++ maltoolbox/language/languagegraph.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 MAL-Toolbox Language Graph Module
 """
 
+import copy
 import logging
 import json
 import zipfile
@@ -380,6 +381,19 @@ class LanguageGraph:
             json.dump(serialized_graph, file, indent=4)
 
 
+    def save_language_specification_to_json(filename: str) -> dict:
+        """
+        Save a MAL language specification dictionary to a JSON file
+
+        Arguments:
+        filename        - the JSON filename where the language specification will be written
+        """
+        logger.info(f'Save language specfication to {filename}.')
+
+        with open(filename, 'w', encoding='utf-8') as file:
+            json.dump(self._lang_spec, file, indent=4)
+
+
     def process_step_expression(self,
         lang: dict,
         target_asset,
@@ -445,9 +459,8 @@ class LanguageGraph:
             case 'variable':
                 # Fetch the step expression associated with the variable from
                 # the language specification and resolve that.
-                variable_step_expr = specification.\
-                    get_variable_for_class_by_name(lang,
-                        target_asset.name, step_expression['name'])
+                variable_step_expr = self._get_variable_for_asset_type_by_name(
+                    target_asset.name, step_expression['name'])
                 if variable_step_expr:
                     return self.process_step_expression(
                         lang,
@@ -525,8 +538,7 @@ class LanguageGraph:
                         dep_chain,
                         step_expression['stepExpression'])
 
-                subtype_asset = next((asset for asset in self.assets \
-                    if asset.name == subtype_name), None)
+                subtype_asset = next((asset for asset in self.assets if asset.name == subtype_name), None)
                 if not subtype_asset:
                     logger.error('Failed to find subtype attack step '
                         f'\"{subtype_name}\"')
@@ -676,9 +688,7 @@ class LanguageGraph:
             logger.debug(f'Create association language graph nodes for asset '
                 f'{asset.name}')
             associations_nodes = []
-            associations = specification.get_associations_for_class(
-                self.lang_dict,
-                asset.name)
+            associations = self._get_associations_for_asset_type(asset.name)
             for association in associations:
                 left_asset = next((asset for asset in self.assets \
                     if asset.name == association['leftAsset']), None)
@@ -738,8 +748,7 @@ class LanguageGraph:
             logger.debug(f'Create attack steps language graph nodes for '
                 f'asset {asset.name}.')
             attack_step_nodes = []
-            attack_steps = specification.get_attacks_for_class(self.lang_dict,
-                asset.name)
+            attack_steps = self._get_attacks_for_asset_type(asset.name)
             for attack_step_name, attack_step_attribs in attack_steps.items():
                 logger.debug(f'Create attack step language graph nodes for '
                     f'{attack_step_name}.')
@@ -813,6 +822,127 @@ class LanguageGraph:
                         [(target_attack_step,
                         self.reverse_dep_chain(dep_chain,
                             None))]
+
+    def _get_attacks_for_asset_type(self, asset_type: str) -> dict:
+        """
+        Get all Attack Steps for a specific Class
+
+        Arguments:
+        asset_type      - a string representing the class for which we want to list
+                          the possible attack steps
+
+        Return:
+        A dictionary representing the set of possible attacks for the specified
+        class. Each key in the dictionary is an attack name and is associated
+        with a dictionary containing other characteristics of the attack such as
+        type of attack, TTC distribution, child attack steps and other information
+        """
+        attack_steps = {}
+        try:
+            asset = next((asset for asset in self._lang_spec['assets'] if asset['name'] == asset_type))
+        except StopIteration:
+            logger.error(f'Failed to find asset type {asset_type} when '\
+                'looking for attack steps.')
+            return None
+
+        logger.debug(f'Get attack steps for {asset["name"]} asset from '\
+            'language specification.')
+        if asset['superAsset']:
+            logger.debug(f'Asset extends another one, fetch the superclass '\
+                'attack steps for it.')
+            attack_steps = self._get_attacks_for_asset_type(asset['superAsset'])
+
+        for step in asset['attackSteps']:
+            if step['name'] not in attack_steps:
+                attack_steps[step['name']] = copy.deepcopy(step)
+            elif not step['reaches']:
+                # This attack step does not lead to any attack steps
+                continue
+            elif step['reaches']['overrides'] == True:
+                attack_steps[step['name']] = copy.deepcopy(step)
+            else:
+                attack_steps[step['name']]['reaches']['stepExpressions'].\
+                    extend(step['reaches']['stepExpressions'])
+
+        return attack_steps
+
+
+    def _get_associations_for_asset_type(self, asset_type: str) -> dict:
+        """
+        Get all Associations for a specific Class
+
+        Arguments:
+        asset_type      - a string representing the class for which we want to list
+                          the associations
+
+        Return:
+        A dictionary representing the set of associations for the specified
+        class. Each key in the dictionary is an attack name and is associated
+        with a dictionary containing other characteristics of the attack such as
+        type of attack, TTC distribution, child attack steps and other information
+        """
+        logger.debug(f'Get associations for {asset_type} asset from '\
+            'language specification.')
+        associations = []
+
+        asset = next((asset for asset in self._lang_spec['assets'] if asset['name'] == \
+            asset_type), None)
+        if not asset:
+            logger.error(f'Failed to find asset type {asset_type} when '\
+                'looking for associations.')
+            return None
+
+        if asset['superAsset']:
+            logger.debug(f'Asset extends another one, fetch the superclass '\
+                'associations for it.')
+            associations.extend(self._get_associations_for_asset_type(asset['superAsset']))
+        assoc_iter = (assoc for assoc in self._lang_spec['associations'] \
+            if assoc['leftAsset'] == asset_type or \
+                assoc['rightAsset'] == asset_type)
+        assoc = next(assoc_iter, None)
+        while (assoc):
+            associations.append(assoc)
+            assoc = next(assoc_iter, None)
+
+        return associations
+
+    def _get_variable_for_asset_type_by_name(self, asset_type: str, variable_name:str) -> dict:
+        """
+        Get a variables for a specific asset type by name.
+        NOTE: Variables are the ones specified in MAL through `let` statements
+
+        Arguments:
+        asset_type      - a string representing the type of asset which contains
+                        the variable
+        variable_name   - the name of the variable to search for
+
+        Return:
+        A dictionary representing the step expressions for the specified variable.
+        """
+
+        asset = next((asset for asset in self._lang_spec['assets'] if asset['name'] == \
+            asset_type), None)
+        if not asset:
+            logger.error(f'Failed to find asset type {asset_type} when '\
+                'looking for variable.')
+            return None
+
+        variable_dict = next((variable for variable in \
+            asset['variables'] if variable['name'] == variable_name), None)
+        if not variable_dict:
+            if asset['superAsset']:
+                variable_dict = self._get_variable_for_asset_type_by_name(asset['superAsset'],
+                                                       variable_name)
+            if variable_dict:
+                return variable_dict
+            else:
+                logger.error(f'Failed to find variable {variable_name} in '\
+                    f'{asset_type}\'s language specification.')
+            return None
+
+        return variable_dict['stepExpression']
+
+
 
     def regenerate_graph(self):
         """
